requirements.txt
langchain
llama-index
langchain_openai
langchain_experimental
langchain_community
neo4j
json-repair
azure-cognitiveservices-vision-computervision
azure-search-documents==11.4.0
azure-functions
azure-identity==1.15.0
azure-core==1.29.5
azure-common==1.1.28
wordcloud~=1.9.3
scikit-learn~=1.4.0
azure-storage-blob==12.19.0
azure-identity
assemblyai~=0.23.1
azure-keyvault-secrets==4.8.0
pandasai~=2.0.37
duckduckgo-search==6.2.1
mutagen==1.47.0
pillow==10.2.0
python-dotenv~=1.0.1
Flask[async]
Flask~=3.0.0
Flask-Migrate==4.0.5
Flask-Navigation==0.2.0
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.6
eventlet~=0.36.1
python-docx==1.1.0
openpyxl==3.1.2
reportlab~=4.1.0
mysql-connector-python~=8.4.0
pyodbc~=5.1.0
msrest~=0.7.1
pymongo~=4.7.2
PyPDF2~=3.0.1
PyMySQL~=1.1.0
docx2txt~=0.8
nltk~=3.8.1
pandas~=1.5.3
matplotlib~=3.8.3
